Add object helper
